api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
python-multipart
requests
SpeechRecognition
pydub
googletrans==4.0.0rc1
python-magic
langdetect